Minor tweaks

+++ src/c_cmds.c
@@ -1777,7 +1777,7 @@ static void condump_cmd_func2(char *cmd, char *parms)
 
     if ((condumpfile = fopen(filename, "wt")))
     {
-        char    *temp = commify(consolestrings - 2);
+        char    *temp = commify((int64_t)consolestrings - 2);
 
         for (int i = 1; i < consolestrings - 1; i++)
             C_DumpConsoleStringToFile(i);
